move Contexts to context folder
