+++ src/Component/Symbolizer/Editor/Editor.tsx
@@ -53,7 +53,7 @@ import { localize } from '../../LocaleWrapper/LocaleWrapper';
 import en_US from '../../../locale/en_US';
 import RasterEditor from '../RasterEditor/RasterEditor';
 import DataUtil from '../../../Util/DataUtil';
-import { CompositionContext, Compositions } from '../../CompositionContext/CompositionContext';
+import { CompositionContext, Compositions } from '../../../context/CompositionContext/CompositionContext';
 import CompositionUtil from '../../../Util/CompositionUtil';
 import { Form } from 'antd';
